@@ -10,34 +10,34 @@
     "start": "next start"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.1.2",
-    "@fortawesome/free-brands-svg-icons": "^6.1.2",
-    "@fortawesome/free-solid-svg-icons": "^6.1.2",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-brands-svg-icons": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@tailwindcss/typography": "^0.5.4",
-    "date-fns": "^2.29.1",
-    "next": "^12.2.3",
+    "@tailwindcss/typography": "^0.5.9",
+    "date-fns": "^2.29.3",
+    "next": "^13.1.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "tailwindcss": "^3.1.7",
-    "unist-util-visit": "^4.1.0"
+    "tailwindcss": "^3.2.4",
+    "unist-util-visit": "^4.1.1"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.8",
-    "eslint": "^8.21.0",
-    "eslint-config-next": "^12.2.3",
-    "github-slugger": "^1.4.0",
-    "globby": "^11.0.0",
+    "autoprefixer": "^10.4.13",
+    "eslint": "^8.31.0",
+    "eslint-config-next": "^13.1.2",
+    "github-slugger": "^2.0.0",
+    "globby": "^13.1.3",
     "gray-matter": "^4.0.3",
-    "moviedb-promise": "^3.1.15",
-    "next-mdx-remote": "^4.1.0",
-    "postcss": "^8.4.14",
-    "prettier": "^2.7.1",
+    "moviedb-promise": "^3.3.1",
+    "next-mdx-remote": "^4.2.1",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.2",
     "rehype-autolink-headings": "^6.1.1",
-    "rehype-highlight": "^5.0.2",
+    "rehype-highlight": "^6.0.0",
     "rehype-katex": "^6.0.2",
     "rehype-raw": "^6.1.1",
-    "rehype-slug": "^5.0.1",
+    "rehype-slug": "^5.1.0",
     "remark-emoji": "^3.0.2",
     "remark-gfm": "^3.0.1",
     "remark-hint": "^1.0.10",
