@@ -8,12 +8,11 @@ export default function AreaLink({ area, category }) {
   return (
     <span className="inline-block lg:text-lg md:text-md sm:text-sm text-gray-500">
       {catLabel} in{" "}
-      <Link href={`/kb/${area}-overview`}>
-        <a
-          className={`inline-block mx-1 px-2 rounded-lg shadow-md ${areaColor} no-underline text-white`}
-        >
-          {emoji.get(emoji_code)} {area}
-        </a>
+      <Link
+        className={`inline-block mx-1 px-2 rounded-lg shadow-md ${areaColor} no-underline text-white`}
+        href={`/kb/${area}-overview`}
+      >
+        {emoji.get(emoji_code)} {area}
       </Link>
     </span>
   )
