@@ -1,4 +1,4 @@
-import globby from "globby"
+import { globby } from "globby"
 import fs from "fs/promises"
 import matter from "gray-matter"
 import GithubSlugger from "github-slugger"
